continue splitting up model loading loging in config into smaller pieces
