+++ app/exceptions.py
@@ -1,0 +1,6 @@
+class ConfigurationException(Exception):
+    pass
+
+
+class ModelLoadingException(Exception):
+    pass
